Workaround pthread dependency behavior change
dependency('threads') used to add the -pthread flags, but it appears to not be doing that at least for C++ files for now. Workaround while following up with the Meson folks is to manualy supply these flags for this dependency. This probably breaks something else, but that is just how it has to be for now.