Convert from variadic template to variadic C-style in log()
Variadic templates can't take advantage of the printf format warning attribute. This causes us to see warnings in forwarded calls. While I would prefer the template approach, I would also like to have proper format warnings enabled and to quash the warning elsewhere.

Clang, at least, is going to allow this, but it is a) non-standard now, and b) not supported in the latest version I can download.

https://reviews.llvm.org/D112579